+++ packages/terra-dynamic-grid/tests/jest/DynamicGrid.test.jsx
@@ -15,8 +15,10 @@ describe('DynamicGrid', () => {
   });
 
   it('should render a Grid with merged attributes', () => {
+    /* eslint-disable react/forbid-component-props */
     const grid = shallow(<DynamicGrid className="TestClass" data-mock="data" style={{ height: '100px' }} />);
     expect(grid).toMatchSnapshot();
+    /* eslint-enable react/forbid-component-props */
   });
 
   // Attributes