+++ packages/terra-form-radio/src/Radio.jsx
@@ -100,7 +100,7 @@ const Radio = ({
   value,
   ...customProps
 }) => {
-  const controlInputAttrs = Object.assign({}, inputAttrs);
+  const controlInputAttrs = { ...inputAttrs };
 
   if (checked !== undefined) {
     controlInputAttrs.checked = checked;