+++ packages/terra-search-field/src/SearchField.jsx
@@ -223,7 +223,7 @@ class SearchField extends React.Component {
     const inputText = intl.formatMessage({ id: 'Terra.searchField.search' });
     const buttonText = intl.formatMessage({ id: 'Terra.searchField.submit-search' });
     const clearText = intl.formatMessage({ id: 'Terra.searchField.clear' });
-    const additionalInputAttributes = Object.assign({}, inputAttributes);
+    const additionalInputAttributes = { 'aria-label': inputText, ...inputAttributes };
     const clearIcon = <span className={cx('clear-icon')} />;
 
     if (value !== undefined) {
@@ -255,7 +255,6 @@ class SearchField extends React.Component {
             onChange={this.handleTextChange}
             disabled={isDisabled}
             aria-disabled={isDisabled}
-            ariaLabel={inputText}
             onKeyDown={this.handleKeyDown}
             refCallback={this.setInputRef}
             {...additionalInputAttributes}